fix warning

+++ src/datawizard/coherency.c
@@ -194,11 +194,11 @@ void _starpu_update_data_state(starpu_data_handle_t handle,
 		unsigned node;
 		for (node = 0; node < nnodes; node++)
 		{
-			if (requesting_replicate->mapped == node
+			if (requesting_replicate->mapped == (int) node
 				&& !_starpu_node_needs_map_update(requesting_node))
 				/* The mapped node will be kept up to date */
 				continue;
-			if (handle->per_node[node].mapped == requesting_node
+			if (handle->per_node[node].mapped == (int) requesting_node
 				&& !_starpu_node_needs_map_update(node))
 				/* The mapping node will be kept up to date */
 				continue;
